dashboard_control: carve api_control.rs (pure move)
Daemon-control requests: settings, displays, voice sessions,
snapshots, log replay, bootstrap, WebRTC signaling, input authority,
worktrees, managed context, MCP tool calls, ControlMsg dispatch, and
the peer/pairing and coordinator surfaces; fifteen tests plus the
stub display backend (struct + impl as a group) ride along.
