+++ src/bin/caller/dashboard_control/api_control.rs
@@ -1,0 +1,2160 @@
+//! Daemon-control requests: settings, displays, voice sessions, state
+//! and browser-workspace snapshots, log replay, dashboard/display
+//! bootstrap, WebRTC signaling, input authority, worktrees, managed
+//! context, MCP tool calls, ControlMsg dispatch, and the peer/pairing
+//! and coordinator surfaces.
+
+use super::*;
+
+pub(crate) async fn api_settings_response(id: String, runtime: &ControlRuntime) -> serde_json::Value {
+    let runtime_settings = {
+        let session = runtime.shared_session.read().await;
+        session.runtime_settings.clone()
+    };
+    json_body_response(
+        id,
+        crate::web_gateway::settings_get_response_body(
+            runtime.project_root.as_deref(),
+            &runtime_settings,
+        )
+        .await,
+        "settings",
+    )
+}
+
+pub(crate) async fn api_displays_response(id: String, runtime: &ControlRuntime) -> serde_json::Value {
+    let session_registry = {
+        let session = runtime.shared_session.read().await;
+        session.session_registry.clone()
+    };
+    json_body_response(
+        id,
+        crate::web_gateway::displays_response_body(&session_registry).await,
+        "displays",
+    )
+}
+
+pub(crate) async fn api_voice_session_response(id: String, runtime: &ControlRuntime) -> serde_json::Value {
+    let provider = runtime
+        .config
+        .get("provider")
+        .and_then(|value| value.as_str())
+        .unwrap_or("gemini");
+    let model = runtime
+        .config
+        .get("model")
+        .and_then(|value| value.as_str())
+        .unwrap_or("");
+    match crate::web_gateway::mint_session_token(provider, model).await {
+        Ok(body) => http_body_response(id, 200, body, "voice session"),
+        Err(msg) => http_body_response(
+            id,
+            502,
+            serde_json::json!({ "error": msg }).to_string(),
+            "voice session",
+        ),
+    }
+}
+
+pub(crate) async fn api_browser_workspace_snapshot_response(id: String) -> serde_json::Value {
+    let workspaces = crate::browser_workspace::list_workspaces().await;
+    serde_json::json!({
+        "t": "response",
+        "id": id,
+        "ok": true,
+        "result": {
+            "t": "browser_workspace_snapshot",
+            "workspaces": workspaces,
+        },
+    })
+}
+
+pub(crate) async fn api_state_snapshot_response(id: String, runtime: &ControlRuntime) -> serde_json::Value {
+    let (daemon_session_id, query_ctx, session_log) = {
+        let session = runtime.shared_session.read().await;
+        (
+            session.daemon_session_id.clone(),
+            session.query_ctx.clone(),
+            session.session_log.clone(),
+        )
+    };
+    let state = query_ctx
+        .as_ref()
+        .map(|ctx| {
+            ctx.agent_state
+                .lock()
+                .unwrap_or_else(|e| e.into_inner())
+                .clone()
+        })
+        .unwrap_or_default();
+    let bootstrap_session_id = daemon_session_id
+        .or_else(|| {
+            query_ctx
+                .as_ref()
+                .and_then(|ctx| control_replay_session_id_from_dir(&ctx.log_dir))
+        })
+        .or_else(|| session_log.as_ref().and_then(control_session_log_id))
+        .unwrap_or_default();
+
+    serde_json::json!({
+        "t": "response",
+        "id": id,
+        "ok": true,
+        "result": {
+            "t": "state_snapshot",
+            "state": state,
+            "connection_id": runtime.session_id.clone(),
+            "config": runtime.config.clone(),
+            "session_id": bootstrap_session_id,
+        },
+    })
+}
+
+pub(crate) async fn api_session_log_replay_response(
+    id: String,
+    runtime: &ControlRuntime,
+) -> serde_json::Value {
+    let replay_log_dir = active_replay_log_dir(runtime).await;
+    let mut replay = replay_log_dir
+        .as_ref()
+        .and_then(|log_dir| {
+            crate::web_gateway::session_log_replay_payload_for_websocket_bootstrap(log_dir)
+        })
+        .and_then(|(payload, external_session_id)| {
+            let mut value = serde_json::from_str::<serde_json::Value>(&payload).ok()?;
+            if let (Some(external_session_id), Some(map)) =
+                (external_session_id, value.as_object_mut())
+            {
+                map.insert(
+                    "external_session_id".to_string(),
+                    serde_json::Value::String(external_session_id),
+                );
+            }
+            Some(value)
+        })
+        .unwrap_or_else(|| {
+            serde_json::json!({
+                "t": "log_replay",
+                "entries": [],
+                "available": false,
+            })
+        });
+    if let Some(map) = replay.as_object_mut() {
+        map.entry("available".to_string())
+            .or_insert(serde_json::Value::Bool(true));
+    }
+
+    serde_json::json!({
+        "t": "response",
+        "id": id,
+        "ok": true,
+        "result": replay,
+    })
+}
+
+pub(crate) async fn api_dashboard_bootstrap_response(
+    id: String,
+    runtime: &ControlRuntime,
+) -> serde_json::Value {
+    let mut frames = Vec::new();
+    if let Some(frame) =
+        response_result(api_state_snapshot_response("bootstrap-state".into(), runtime).await)
+    {
+        frames.push(frame);
+    }
+    if let Some(result) = response_result(cached_bootstrap_events_response_frame(
+        "bootstrap-cached".into(),
+        &runtime.bootstrap_caches,
+    )) {
+        if let Some(events) = result.get("events").and_then(|value| value.as_array()) {
+            frames.extend(events.iter().cloned());
+        }
+    }
+    if let Some(frame) =
+        response_result(api_browser_workspace_snapshot_response("bootstrap-browser".into()).await)
+    {
+        frames.push(frame);
+    }
+    frames.extend(display_ready_bootstrap_frames(runtime).await);
+    let mut replayed_external_session_ids = HashSet::new();
+    if let Some(frame) =
+        response_result(api_session_log_replay_response("bootstrap-replay".into(), runtime).await)
+    {
+        if let Some(external_session_id) = frame
+            .get("external_session_id")
+            .and_then(|value| value.as_str())
+        {
+            replayed_external_session_ids.insert(external_session_id.to_string());
+        }
+        frames.push(frame);
+    }
+    frames.extend(external_session_activity_replay_frames(
+        runtime,
+        &replayed_external_session_ids,
+    ));
+    frames.extend(display_authority_snapshot_frames(runtime).await);
+    let frame_count = frames.len();
+    let omitted = dashboard_bootstrap_omitted(runtime);
+
+    serde_json::json!({
+        "t": "response",
+        "id": id,
+        "ok": true,
+        "result": {
+            "frames": frames,
+            "frame_count": frame_count,
+            "omitted": omitted,
+        },
+    })
+}
+
+pub(crate) async fn api_display_bootstrap_response(id: String, runtime: &ControlRuntime) -> serde_json::Value {
+    let mut frames = display_ready_bootstrap_frames(runtime).await;
+    frames.extend(display_authority_snapshot_frames(runtime).await);
+    let frame_count = frames.len();
+    let omitted = display_bootstrap_omitted(runtime);
+    serde_json::json!({
+        "t": "response",
+        "id": id,
+        "ok": true,
+        "result": {
+            "frames": frames,
+            "frame_count": frame_count,
+            "omitted": omitted,
+        },
+    })
+}
+
+pub(crate) async fn api_display_webrtc_signal_response(
+    id: String,
+    params: Option<&serde_json::Value>,
+    runtime: &ControlRuntime,
+) -> serde_json::Value {
+    let params = params.cloned().unwrap_or_else(|| serde_json::json!({}));
+    let signal = string_param(&params, &["signal", "kind", "type", "t"]);
+    match signal.as_str() {
+        "offer" | "display_offer" => api_display_webrtc_offer_response(id, &params, runtime).await,
+        "ice" | "candidate" | "display_ice" => {
+            api_display_webrtc_ice_response(id, &params, runtime).await
+        }
+        _ => serde_json::json!({
+            "t": "response",
+            "id": id,
+            "ok": false,
+            "error": "missing or unknown display webrtc signal",
+        }),
+    }
+}
+
+pub(crate) async fn api_display_webrtc_offer_response(
+    id: String,
+    params: &serde_json::Value,
+    runtime: &ControlRuntime,
+) -> serde_json::Value {
+    let display_id = display_id_param(Some(params));
+    let sdp = string_param(params, &["sdp", "offer", "offer_sdp"]);
+    if sdp.is_empty() {
+        return missing_param_response(id, "sdp");
+    }
+    let Some(display_session) = active_display_session(runtime, display_id).await else {
+        return display_signal_error_response(id, 404, display_id, "display session not found");
+    };
+
+    let (ice_tx, mut ice_rx) = mpsc::channel::<(crate::display::PeerId, String)>(64);
+    if let Some(control_frames_tx) = runtime.control_frames_tx.clone() {
+        tokio::spawn(async move {
+            while let Some((_peer_id, candidate_json)) = ice_rx.recv().await {
+                let candidate =
+                    serde_json::from_str::<serde_json::Value>(&candidate_json).unwrap_or_default();
+                let payload = serde_json::json!({
+                    "t": "display_ice",
+                    "display_id": display_id,
+                    "candidate": candidate,
+                });
+                let frame = serde_json::json!({
+                    "t": "event",
+                    "payload": payload,
+                });
+                if control_frames_tx.send(frame).is_err() {
+                    break;
+                }
+            }
+        });
+    }
+
+    let input_authorized = dashboard_display_input_authorizer(
+        runtime.display_authority.clone(),
+        runtime.session_id.clone(),
+        display_id,
+    );
+    let authority_handler = crate::display::webrtc::noop_authority_handler();
+    match display_session
+        .handle_offer(
+            runtime.display_peer_id,
+            &sdp,
+            &runtime.ice_config,
+            Some(Arc::clone(&runtime.tcp_peer_registry)),
+            runtime.tcp_advertised,
+            ice_tx,
+            input_authorized,
+            authority_handler,
+        )
+        .await
+    {
+        Ok(answer_sdp) => serde_json::json!({
+            "t": "response",
+            "id": id,
+            "ok": true,
+            "result": {
+                "t": "display_answer",
+                "display_id": display_id,
+                "sdp": answer_sdp,
+            },
+        }),
+        Err(e) => display_signal_error_response(
+            id,
+            502,
+            display_id,
+            &format!("display offer failed: {e}"),
+        ),
+    }
+}
+
+pub(crate) async fn api_display_webrtc_ice_response(
+    id: String,
+    params: &serde_json::Value,
+    runtime: &ControlRuntime,
+) -> serde_json::Value {
+    let display_id = display_id_param(Some(params));
+    let Some(candidate) = params.get("candidate").cloned() else {
+        return missing_param_response(id, "candidate");
+    };
+    if candidate.is_null() {
+        return missing_param_response(id, "candidate");
+    }
+    let Some(display_session) = active_display_session(runtime, display_id).await else {
+        return display_signal_error_response(id, 404, display_id, "display session not found");
+    };
+    let candidate = candidate.to_string();
+    let peer_id = runtime.display_peer_id;
+    tokio::spawn(async move {
+        if let Err(e) = display_session.add_ice_candidate(peer_id, &candidate).await {
+            eprintln!(
+                "[dashboard/control] display ICE candidate failed for display {display_id}: {e}"
+            );
+        }
+    });
+    serde_json::json!({
+        "t": "response",
+        "id": id,
+        "ok": true,
+        "result": {
+            "ok": true,
+            "display_id": display_id,
+        },
+    })
+}
+
+pub(crate) async fn active_display_session(
+    runtime: &ControlRuntime,
+    display_id: u32,
+) -> Option<Arc<crate::display::DisplaySession>> {
+    let session_registry = {
+        let session = runtime.shared_session.read().await;
+        session.session_registry.clone()
+    }?;
+    let registry = session_registry.read().await;
+    registry.get(display_id)
+}
+
+pub(crate) fn dashboard_display_input_authorizer(
+    display_authority: Option<DashboardDisplayAuthorityBridge>,
+    session_id: String,
+    display_id: u32,
+) -> Arc<dyn Fn() -> bool + Send + Sync> {
+    Arc::new(move || match display_authority.as_ref() {
+        Some(bridge) => bridge.input_authorized(&session_id, display_id),
+        None => true,
+    })
+}
+
+pub(crate) fn display_signal_error_response(
+    id: String,
+    status: u16,
+    display_id: u32,
+    error: &str,
+) -> serde_json::Value {
+    serde_json::json!({
+        "t": "response",
+        "id": id,
+        "ok": false,
+        "status": status,
+        "display_id": display_id,
+        "error": error,
+    })
+}
+
+pub(crate) async fn api_display_input_authority_snapshot_response(
+    id: String,
+    runtime: &ControlRuntime,
+) -> serde_json::Value {
+    let frames = display_authority_snapshot_frames(runtime).await;
+    let frame_count = frames.len();
+    serde_json::json!({
+        "t": "response",
+        "id": id,
+        "ok": true,
+        "result": {
+            "available": runtime.display_authority.is_some(),
+            "frames": frames,
+            "frame_count": frame_count,
+        },
+    })
+}
+
+pub(crate) async fn api_display_input_authority_request_response(
+    id: String,
+    params: Option<&serde_json::Value>,
+    runtime: &ControlRuntime,
+) -> serde_json::Value {
+    let Some(bridge) = runtime.display_authority.as_ref() else {
+        return display_authority_unavailable_response(id);
+    };
+    let display_id = display_id_param(params);
+    let frames = bridge.request(&runtime.session_id, display_id);
+    let frame_count = frames.len();
+    serde_json::json!({
+        "t": "response",
+        "id": id,
+        "ok": true,
+        "result": {
+            "ok": true,
+            "display_id": display_id,
+            "frames": frames,
+            "frame_count": frame_count,
+        },
+    })
+}
+
+pub(crate) async fn api_display_input_authority_release_response(
+    id: String,
+    params: Option<&serde_json::Value>,
+    runtime: &ControlRuntime,
+) -> serde_json::Value {
+    let Some(bridge) = runtime.display_authority.as_ref() else {
+        return display_authority_unavailable_response(id);
+    };
+    let display_id = display_id_param(params);
+    let frames = bridge.release(&runtime.session_id, display_id);
+    let frame_count = frames.len();
+    serde_json::json!({
+        "t": "response",
+        "id": id,
+        "ok": true,
+        "result": {
+            "ok": true,
+            "display_id": display_id,
+            "frames": frames,
+            "frame_count": frame_count,
+        },
+    })
+}
+
+pub(crate) fn display_authority_unavailable_response(id: String) -> serde_json::Value {
+    serde_json::json!({
+        "t": "response",
+        "id": id,
+        "ok": true,
+        "result": {
+            "ok": false,
+            "available": false,
+            "_httpStatus": 503,
+            "_httpOk": false,
+            "error": "display input authority unavailable",
+        },
+    })
+}
+
+pub(crate) fn display_id_param(params: Option<&serde_json::Value>) -> u32 {
+    params
+        .and_then(|params| {
+            params
+                .get("display_id")
+                .or_else(|| params.get("displayId"))
+                .or_else(|| params.get("id"))
+        })
+        .and_then(|value| value.as_u64())
+        .and_then(|value| u32::try_from(value).ok())
+        .unwrap_or(0)
+}
+
+pub(crate) async fn display_authority_snapshot_frames(runtime: &ControlRuntime) -> Vec<serde_json::Value> {
+    let Some(bridge) = runtime.display_authority.as_ref() else {
+        return Vec::new();
+    };
+    let display_ids = active_display_ids(runtime).await;
+    bridge.snapshot(&runtime.session_id, &display_ids)
+}
+
+pub(crate) fn dashboard_bootstrap_omitted(runtime: &ControlRuntime) -> Vec<&'static str> {
+    if runtime.display_authority.is_some() {
+        Vec::new()
+    } else {
+        vec!["display_input_authority_state"]
+    }
+}
+
+pub(crate) fn display_bootstrap_omitted(runtime: &ControlRuntime) -> Vec<&'static str> {
+    if runtime.display_authority.is_some() {
+        Vec::new()
+    } else {
+        vec!["display_input_authority_state"]
+    }
+}
+
+pub(crate) async fn display_ready_bootstrap_frames(runtime: &ControlRuntime) -> Vec<serde_json::Value> {
+    let display_ids = active_display_ids(runtime).await;
+    let session_registry = {
+        let session = runtime.shared_session.read().await;
+        session.session_registry.clone()
+    };
+    let Some(session_registry) = session_registry else {
+        return Vec::new();
+    };
+
+    let registry = session_registry.read().await;
+    display_ids
+        .into_iter()
+        .filter_map(|display_id| {
+            registry.get(display_id).map(|session| {
+                let (width, height) = session.resolution();
+                serde_json::json!({
+                    "event": "display_ready",
+                    "display_id": display_id,
+                    "width": width,
+                    "height": height,
+                })
+            })
+        })
+        .collect()
+}
+
+pub(crate) async fn active_display_ids(runtime: &ControlRuntime) -> Vec<u32> {
+    let session_registry = {
+        let session = runtime.shared_session.read().await;
+        session.session_registry.clone()
+    };
+    let Some(session_registry) = session_registry else {
+        return Vec::new();
+    };
+
+    let registry = session_registry.read().await;
+    let mut display_ids = registry.display_ids();
+    display_ids.sort_unstable();
+    display_ids
+}
+
+pub(crate) async fn api_external_session_activity_replay_response(
+    id: String,
+    runtime: &ControlRuntime,
+) -> serde_json::Value {
+    let frames = external_session_activity_replay_frames(runtime, &HashSet::new());
+    let frame_count = frames.len();
+    serde_json::json!({
+        "t": "response",
+        "id": id,
+        "ok": true,
+        "result": {
+            "frames": frames,
+            "frame_count": frame_count,
+        },
+    })
+}
+
+pub(crate) fn external_session_activity_replay_frames(
+    runtime: &ControlRuntime,
+    skip_session_ids: &HashSet<String>,
+) -> Vec<serde_json::Value> {
+    let mut active_external_sessions: Vec<(String, String)> = runtime
+        .bootstrap_caches
+        .attached_external_sessions
+        .lock()
+        .ok()
+        .map(|guard| {
+            guard
+                .iter()
+                .map(|(session_id, source)| (session_id.clone(), source.clone()))
+                .collect()
+        })
+        .unwrap_or_default();
+    active_external_sessions.sort_by(|a, b| a.0.cmp(&b.0));
+    active_external_sessions
+        .into_iter()
+        .filter(|(session_id, _)| !skip_session_ids.contains(session_id))
+        .filter_map(|(session_id, source)| {
+            crate::web_gateway::external_session_activity_replay_for_websocket(&source, &session_id)
+                .and_then(|payload| serde_json::from_str::<serde_json::Value>(&payload).ok())
+        })
+        .collect()
+}
+
+pub(crate) fn response_result(response: serde_json::Value) -> Option<serde_json::Value> {
+    response.get("result").cloned()
+}
+
+pub(crate) fn control_replay_session_id_from_dir(log_dir: &std::path::Path) -> Option<String> {
+    std::fs::read_to_string(log_dir.join("session_meta.json"))
+        .ok()
+        .and_then(|meta| serde_json::from_str::<crate::session_log::SessionMeta>(&meta).ok())
+        .map(|meta| meta.session_id)
+        .filter(|session_id| !session_id.trim().is_empty())
+        .or_else(|| {
+            log_dir
+                .file_name()
+                .map(|name| name.to_string_lossy().to_string())
+                .filter(|session_id| !session_id.trim().is_empty())
+        })
+}
+
+pub(crate) fn control_session_log_id(
+    session_log: &Arc<std::sync::Mutex<crate::session_log::SessionLog>>,
+) -> Option<String> {
+    session_log
+        .lock()
+        .ok()
+        .map(|log| log.session_id().to_string())
+        .filter(|id| !id.trim().is_empty())
+}
+
+pub(crate) async fn active_replay_log_dir(runtime: &ControlRuntime) -> Option<PathBuf> {
+    let (query_ctx, session_log) = {
+        let session = runtime.shared_session.read().await;
+        (session.query_ctx.clone(), session.session_log.clone())
+    };
+    query_ctx
+        .as_ref()
+        .map(|ctx| ctx.log_dir.clone())
+        .or_else(|| {
+            session_log
+                .as_ref()
+                .and_then(|log| log.lock().ok().map(|log| log.dir().to_path_buf()))
+        })
+}
+
+pub(crate) async fn api_worktrees_response(id: String, runtime: &ControlRuntime) -> serde_json::Value {
+    let body = runtime
+        .worktree_inventory_cache
+        .lock()
+        .ok()
+        .and_then(|guard| guard.clone())
+        .unwrap_or_else(crate::web_gateway::empty_worktree_inventory_response);
+    json_body_response(id, body, "worktrees")
+}
+
+pub(crate) async fn api_worktrees_inspect_response(
+    id: String,
+    params: Option<&serde_json::Value>,
+    _runtime: &ControlRuntime,
+) -> serde_json::Value {
+    let body_text = params_body_text(params);
+    let home = dirs::home_dir().unwrap_or_else(|| PathBuf::from("/tmp"));
+    let result = tokio::task::spawn_blocking(move || {
+        crate::web_gateway::inspect_worktree_inventory_response(&home, &body_text)
+    })
+    .await;
+    match result {
+        Ok((status_line, body)) => {
+            http_body_response(id, status_line_code(status_line), body, "worktree inspect")
+        }
+        Err(e) => http_body_response(
+            id,
+            500,
+            serde_json::json!({
+                "ok": false,
+                "error": format!("worktree inspect task failed: {e}")
+            })
+            .to_string(),
+            "worktree inspect",
+        ),
+    }
+}
+
+pub(crate) async fn api_worktrees_scan_response(id: String, runtime: &ControlRuntime) -> serde_json::Value {
+    let home = dirs::home_dir().unwrap_or_else(|| PathBuf::from("/tmp"));
+    let project_root = runtime.project_root.clone();
+    let cache = runtime.worktree_inventory_cache.clone();
+    let result = tokio::task::spawn_blocking(move || {
+        let body =
+            crate::web_gateway::scan_worktree_inventory_response(&home, project_root.as_deref());
+        if let Ok(mut guard) = cache.lock() {
+            *guard = Some(body.clone());
+        }
+        body
+    })
+    .await;
+    match result {
+        Ok(body) => json_body_response(id, body, "worktree scan"),
+        Err(e) => http_body_response(
+            id,
+            500,
+            serde_json::json!({
+                "error": format!("worktree scan task failed: {e}")
+            })
+            .to_string(),
+            "worktree scan",
+        ),
+    }
+}
+
+pub(crate) async fn api_worktrees_remove_response(
+    id: String,
+    params: Option<&serde_json::Value>,
+    runtime: &ControlRuntime,
+) -> serde_json::Value {
+    let body_text = params_body_text(params);
+    let home = dirs::home_dir().unwrap_or_else(|| PathBuf::from("/tmp"));
+    let cache = runtime.worktree_inventory_cache.clone();
+    let result = tokio::task::spawn_blocking(move || {
+        let result = crate::web_gateway::remove_worktree_inventory_response(&home, &body_text);
+        if result.0 == "200 OK" {
+            if let Ok(mut guard) = cache.lock() {
+                *guard = None;
+            }
+        }
+        result
+    })
+    .await;
+    match result {
+        Ok((status_line, body)) => {
+            http_body_response(id, status_line_code(status_line), body, "worktree remove")
+        }
+        Err(e) => http_body_response(
+            id,
+            500,
+            serde_json::json!({
+                "ok": false,
+                "error": format!("worktree removal task failed: {e}")
+            })
+            .to_string(),
+            "worktree remove",
+        ),
+    }
+}
+
+pub(crate) async fn api_managed_context_response(
+    id: String,
+    kind: &'static str,
+    params: Option<&serde_json::Value>,
+    runtime: &ControlRuntime,
+) -> serde_json::Value {
+    let params = params.cloned().unwrap_or_else(|| serde_json::json!({}));
+    let Some(request_line) = managed_context_request_line(kind, &params) else {
+        return missing_param_response(id, "query");
+    };
+    let active_log_dir = match active_session_log_dir(runtime).await {
+        Ok(dir) => dir,
+        Err(error) => {
+            return http_body_response(
+                id,
+                500,
+                serde_json::json!({ "error": error }).to_string(),
+                "managed context",
+            );
+        }
+    };
+    let home = crate::platform::home_dir();
+    let response = tokio::task::spawn_blocking(move || match kind {
+        "records" => crate::web_gateway::managed_context_records_response_from_home(
+            &request_line,
+            active_log_dir.as_deref(),
+            &home,
+        ),
+        "anchors" => crate::web_gateway::managed_context_anchors_response_from_home(
+            &request_line,
+            active_log_dir.as_deref(),
+            &home,
+        ),
+        "fission" => crate::web_gateway::managed_context_fission_response_from_home(
+            &request_line,
+            active_log_dir.as_deref(),
+            &home,
+        ),
+        _ => crate::web_gateway::managed_context_records_response_from_home(
+            &request_line,
+            active_log_dir.as_deref(),
+            &home,
+        ),
+    })
+    .await;
+    let response = match response {
+        Ok(response) => response,
+        Err(e) => {
+            return serde_json::json!({
+                "t": "response",
+                "id": id,
+                "ok": false,
+                "error": format!("managed context task failed: {e}"),
+            });
+        }
+    };
+    http_wire_response(id, response, "managed context")
+}
+
+pub(crate) async fn api_mcp_tool_call_response(
+    id: String,
+    params: Option<&serde_json::Value>,
+    runtime: &ControlRuntime,
+) -> serde_json::Value {
+    let params = params.cloned().unwrap_or_else(|| serde_json::json!({}));
+    let mcp_id = params
+        .get("mcp_id")
+        .or_else(|| params.get("rpc_id"))
+        .cloned()
+        .unwrap_or_else(|| serde_json::json!(id.clone()));
+    let Some(server) = runtime.mcp_server.as_ref() else {
+        return http_body_response(
+            id,
+            503,
+            mcp_error_body(mcp_id, -32603, "MCP server not available"),
+            "mcp tool call",
+        );
+    };
+    let session_id = optional_string_param(
+        &params,
+        &["session_id", "session", "intendant_session", "sessionId"],
+    );
+    if session_id.is_none() {
+        return http_body_response(
+            id,
+            400,
+            mcp_error_body(mcp_id, -32602, "missing session_id"),
+            "mcp tool call",
+        );
+    }
+    let name = string_param(&params, &["name", "tool", "tool_name"]);
+    if name.is_empty() {
+        return http_body_response(
+            id,
+            400,
+            mcp_error_body(mcp_id, -32602, "missing tool name"),
+            "mcp tool call",
+        );
+    }
+    // Layered on top of the dispatch-level `message.send` gate: the named
+    // tool must also clear its own IAM operation, so a principal scoped to
+    // messaging cannot reach display input or runtime control through the
+    // generic tool-call RPC.
+    let decision = runtime
+        .grant
+        .access_decision(crate::mcp::mcp_tool_operation(&name));
+    if !decision.allowed {
+        return http_body_response(
+            id,
+            403,
+            mcp_error_body(
+                mcp_id,
+                -32603,
+                &format!(
+                    "permission denied for tool '{name}': {} (permission {})",
+                    decision.reason, decision.permission
+                ),
+            ),
+            "mcp tool call",
+        );
+    }
+    let arguments = params
+        .get("arguments")
+        .or_else(|| params.get("args"))
+        .cloned()
+        .unwrap_or_else(|| serde_json::json!({}));
+    let managed_context = optional_managed_context_param(&params);
+    match server
+        .call_tool_by_name_for_session(&name, arguments, session_id.as_deref(), managed_context)
+        .await
+    {
+        Ok(result) => {
+            let result = serde_json::to_value(result).unwrap_or_else(|e| {
+                serde_json::json!({
+                    "content": [{
+                        "type": "text",
+                        "text": format!("Failed to serialize MCP tool result: {}", e),
+                    }],
+                    "isError": true,
+                })
+            });
+            http_body_response(
+                id,
+                200,
+                serde_json::json!({
+                    "jsonrpc": "2.0",
+                    "id": mcp_id,
+                    "result": result,
+                })
+                .to_string(),
+                "mcp tool call",
+            )
+        }
+        Err(error) => http_body_response(
+            id,
+            200,
+            mcp_error_body(mcp_id, -32603, &error),
+            "mcp tool call",
+        ),
+    }
+}
+
+pub(crate) fn mcp_error_body(id: serde_json::Value, code: i64, message: &str) -> String {
+    serde_json::json!({
+        "jsonrpc": "2.0",
+        "id": id,
+        "error": {
+            "code": code,
+            "message": message,
+        },
+    })
+    .to_string()
+}
+
+pub(crate) fn optional_managed_context_param(params: &serde_json::Value) -> Option<bool> {
+    for name in ["managed_context", "managedContext", "codex_managed_context"] {
+        let Some(value) = params.get(name) else {
+            continue;
+        };
+        if let Some(flag) = value.as_bool() {
+            return Some(flag);
+        }
+        if let Some(mode) = value.as_str() {
+            return Some(crate::project::codex_managed_context_enabled(mode));
+        }
+    }
+    None
+}
+
+pub(crate) async fn api_settings_save_response(
+    id: String,
+    params: Option<&serde_json::Value>,
+    runtime: &ControlRuntime,
+) -> serde_json::Value {
+    let body_text = params_body_text(params);
+    let (status_line, body) = crate::web_gateway::settings_post_result(
+        &body_text,
+        runtime.project_root.as_deref(),
+        &runtime.bus,
+    );
+    http_body_response(id, status_line_code(status_line), body, "settings save")
+}
+
+pub(crate) async fn api_control_msg_response(
+    id: String,
+    params: Option<&serde_json::Value>,
+    runtime: &ControlRuntime,
+) -> serde_json::Value {
+    let ctrl = match dashboard_control_msg_from_params(id.clone(), params) {
+        Ok(ctrl) => ctrl,
+        Err(response) => return response,
+    };
+    if !dashboard_control_msg_allowed(&ctrl) {
+        return http_body_response(
+            id,
+            400,
+            serde_json::json!({
+                "ok": false,
+                "error": format!(
+                    "control action not available over dashboard WebRTC: {}",
+                    dashboard_control_msg_action(&ctrl)
+                ),
+            })
+            .to_string(),
+            "control message",
+        );
+    }
+    let action = dashboard_control_msg_action(&ctrl);
+    runtime.bus.send(AppEvent::PresenceLog {
+        message: format!("[dashboard-control] ControlMsg: {action}"),
+        level: Some(crate::types::LogLevel::Debug),
+        turn: None,
+    });
+    runtime.bus.send(AppEvent::ControlCommand(ctrl));
+    serde_json::json!({
+        "t": "response",
+        "id": id,
+        "ok": true,
+        "result": {
+            "ok": true,
+            "action": action,
+        },
+    })
+}
+
+pub(crate) async fn api_session_control_msg_response(
+    id: String,
+    params: Option<&serde_json::Value>,
+    runtime: &ControlRuntime,
+) -> serde_json::Value {
+    let ctrl = match dashboard_control_msg_from_params(id.clone(), params) {
+        Ok(ctrl) => ctrl,
+        Err(response) => return response,
+    };
+    if !dashboard_session_control_msg_allowed(&ctrl) {
+        return http_body_response(
+            id,
+            400,
+            serde_json::json!({
+                "ok": false,
+                "error": format!(
+                    "control action not available over dashboard session WebRTC: {}",
+                    dashboard_control_msg_action(&ctrl)
+                ),
+            })
+            .to_string(),
+            "session control message",
+        );
+    }
+    let action = dashboard_control_msg_action(&ctrl);
+    dispatch_dashboard_control_msg(&runtime.bus, ctrl, "session-control");
+    serde_json::json!({
+        "t": "response",
+        "id": id,
+        "ok": true,
+        "result": {
+            "ok": true,
+            "action": action,
+        },
+    })
+}
+
+pub(crate) async fn api_dashboard_action_msg_response(
+    id: String,
+    params: Option<&serde_json::Value>,
+    runtime: &ControlRuntime,
+) -> serde_json::Value {
+    let ctrl = match dashboard_control_msg_from_params(id.clone(), params) {
+        Ok(ctrl) => ctrl,
+        Err(response) => return response,
+    };
+    if !dashboard_action_msg_allowed(&ctrl) {
+        return http_body_response(
+            id,
+            400,
+            serde_json::json!({
+                "ok": false,
+                "error": format!(
+                    "control action not available over dashboard action WebRTC: {}",
+                    dashboard_control_msg_action(&ctrl)
+                ),
+            })
+            .to_string(),
+            "dashboard action message",
+        );
+    }
+    let action = dashboard_control_msg_action(&ctrl);
+    let marker_apply = match &ctrl {
+        ControlMsg::SetDiagnosticsVisualMarker {
+            display_id,
+            enabled,
+        } => {
+            let display_id = display_id.unwrap_or(0);
+            Some((
+                display_id,
+                apply_dashboard_diagnostics_visual_marker(runtime, display_id, *enabled).await,
+            ))
+        }
+        _ => None,
+    };
+    dispatch_dashboard_control_msg(&runtime.bus, ctrl, "dashboard-action");
+    let mut result = serde_json::json!({
+        "ok": true,
+        "action": action,
+    });
+    if let Some((display_id, marker_result)) = marker_apply {
+        if let Some(result_obj) = result.as_object_mut() {
+            result_obj.insert("display_id".to_string(), serde_json::json!(display_id));
+            result_obj.insert(
+                "registry_available".to_string(),
+                serde_json::json!(marker_result.registry_available),
+            );
+            result_obj.insert(
+                "active_display_updated".to_string(),
+                serde_json::json!(marker_result.active_display_updated),
+            );
+        }
+    }
+    serde_json::json!({
+        "t": "response",
+        "id": id,
+        "ok": true,
+        "result": result,
+    })
+}
+
+pub(crate) async fn api_diagnostics_visual_freshness_response(
+    id: String,
+    params: Option<&serde_json::Value>,
+) -> serde_json::Value {
+    let params = params.cloned().unwrap_or_else(|| serde_json::json!({}));
+    let session_id = string_param(&params, &["session_id", "sessionId", "id"]);
+    if session_id.is_empty() {
+        return missing_param_response(id, "session_id");
+    }
+    let body = params
+        .get("body")
+        .or_else(|| params.get("ndjson"))
+        .and_then(|value| value.as_str())
+        .map(ToString::to_string)
+        .unwrap_or_default();
+    if body.is_empty() {
+        return missing_param_response(id, "body");
+    }
+
+    let result = tokio::task::spawn_blocking(move || {
+        crate::diagnostics::append_visual_freshness_record(&session_id, body.as_bytes())
+    })
+    .await;
+    let (status, body) = match result {
+        Ok(Ok(written)) => (
+            200,
+            serde_json::json!({"ok": true, "written": written}).to_string(),
+        ),
+        Ok(Err(e)) if e.kind() == std::io::ErrorKind::InvalidInput => {
+            (400, serde_json::json!({"error": e.to_string()}).to_string())
+        }
+        Ok(Err(e)) => (500, serde_json::json!({"error": e.to_string()}).to_string()),
+        Err(e) => (
+            500,
+            serde_json::json!({"error": format!("diagnostics append task failed: {e}")})
+                .to_string(),
+        ),
+    };
+    http_body_response(id, status, body, "diagnostics visual freshness")
+}
+
+pub(crate) fn dashboard_control_msg_from_params(
+    id: String,
+    params: Option<&serde_json::Value>,
+) -> Result<ControlMsg, serde_json::Value> {
+    let Some(params) = params else {
+        return Err(missing_param_response(id, "message"));
+    };
+    let message = params
+        .get("message")
+        .or_else(|| params.get("control_msg"))
+        .or_else(|| params.get("controlMsg"))
+        .cloned()
+        .unwrap_or_else(|| params.clone());
+    serde_json::from_value::<ControlMsg>(message).map_err(|e| {
+        http_body_response(
+            id,
+            400,
+            serde_json::json!({
+                "ok": false,
+                "error": format!("invalid control message: {e}"),
+            })
+            .to_string(),
+            "control message",
+        )
+    })
+}
+
+pub(crate) fn dispatch_dashboard_control_msg(bus: &crate::event::EventBus, ctrl: ControlMsg, scope: &str) {
+    let action = dashboard_control_msg_action(&ctrl);
+    bus.send(AppEvent::PresenceLog {
+        message: format!("[dashboard-control:{scope}] ControlMsg: {action}"),
+        level: Some(crate::types::LogLevel::Debug),
+        turn: None,
+    });
+    bus.send(AppEvent::ControlCommand(ctrl));
+}
+
+#[derive(Debug, Clone, Copy)]
+pub(crate) struct DiagnosticsVisualMarkerApply {
+    registry_available: bool,
+    active_display_updated: bool,
+}
+
+pub(crate) async fn apply_dashboard_diagnostics_visual_marker(
+    runtime: &ControlRuntime,
+    display_id: u32,
+    enabled: bool,
+) -> DiagnosticsVisualMarkerApply {
+    let session_registry = {
+        let session = runtime.shared_session.read().await;
+        session.session_registry.clone()
+    };
+    let Some(session_registry) = session_registry else {
+        eprintln!(
+            "[dashboard/control] diagnostics visual marker request for display {display_id} ({enabled}) ignored; no session registry"
+        );
+        return DiagnosticsVisualMarkerApply {
+            registry_available: false,
+            active_display_updated: false,
+        };
+    };
+
+    let active_display_updated = session_registry
+        .write()
+        .await
+        .set_diagnostics_visual_marker(display_id, enabled);
+    eprintln!(
+        "[dashboard/control] diagnostics visual marker for display {display_id} = {enabled}{}",
+        if active_display_updated {
+            ""
+        } else {
+            " (pending)"
+        },
+    );
+    DiagnosticsVisualMarkerApply {
+        registry_available: true,
+        active_display_updated,
+    }
+}
+
+pub(crate) fn dashboard_control_msg_allowed(ctrl: &ControlMsg) -> bool {
+    matches!(
+        ctrl,
+        ControlMsg::SetAutonomy { .. }
+            | ControlMsg::SetApprovalRule { .. }
+            | ControlMsg::SetExternalAgent { .. }
+            | ControlMsg::SetCodexCommand { .. }
+            | ControlMsg::SetCodexManagedCommand { .. }
+            | ControlMsg::SetCodexSandbox { .. }
+            | ControlMsg::SetCodexApprovalPolicy { .. }
+            | ControlMsg::SetCodexModel { .. }
+            | ControlMsg::SetCodexReasoningEffort { .. }
+            | ControlMsg::SetCodexServiceTier { .. }
+            | ControlMsg::SetCodexWebSearch { .. }
+            | ControlMsg::SetCodexNetworkAccess { .. }
+            | ControlMsg::SetCodexWritableRoots { .. }
+            | ControlMsg::SetCodexManagedContext { .. }
+            | ControlMsg::SetCodexContextArchive { .. }
+            | ControlMsg::SetClaudeModel { .. }
+            | ControlMsg::SetClaudePermissionMode { .. }
+            | ControlMsg::SetClaudeAllowedTools { .. }
+            | ControlMsg::SetVerbosity { .. }
+    )
+}
+
+pub(crate) fn dashboard_session_control_msg_allowed(ctrl: &ControlMsg) -> bool {
+    matches!(
+        ctrl,
+        ControlMsg::Approve { .. }
+            | ControlMsg::Deny { .. }
+            | ControlMsg::Skip { .. }
+            | ControlMsg::ApproveAll { .. }
+            | ControlMsg::AnswerQuestion { .. }
+            | ControlMsg::RenameSession { .. }
+            | ControlMsg::ConfigureSessionAgent { .. }
+            | ControlMsg::StopSession { .. }
+            | ControlMsg::RestartSession { .. }
+            | ControlMsg::CreateSession { .. }
+            | ControlMsg::SpawnSubAgent { .. }
+            | ControlMsg::StartTask { .. }
+            | ControlMsg::ResumeSession { .. }
+            | ControlMsg::FollowUp { .. }
+            | ControlMsg::CancelFollowUp { .. }
+            | ControlMsg::EditUserMessage { .. }
+            | ControlMsg::Interrupt { .. }
+            | ControlMsg::Steer { .. }
+            | ControlMsg::CancelSteer { .. }
+    )
+}
+
+pub(crate) fn dashboard_action_msg_allowed(ctrl: &ControlMsg) -> bool {
+    matches!(
+        ctrl,
+        ControlMsg::CodexThreadAction { .. }
+            | ControlMsg::TakeDisplay { .. }
+            | ControlMsg::ReleaseDisplay { .. }
+            | ControlMsg::GrantUserDisplay { .. }
+            | ControlMsg::RevokeUserDisplay { .. }
+            | ControlMsg::CreateBrowserWorkspace { .. }
+            | ControlMsg::CloseBrowserWorkspace { .. }
+            | ControlMsg::AcquireBrowserWorkspace { .. }
+            | ControlMsg::ReleaseBrowserWorkspace { .. }
+            | ControlMsg::SetupDebugScreen
+            | ControlMsg::TeardownDebugScreen
+            | ControlMsg::StartDebugRecording
+            | ControlMsg::StopDebugRecording
+            | ControlMsg::StartRecording { .. }
+            | ControlMsg::StopRecording { .. }
+            | ControlMsg::DeleteRecording { .. }
+            | ControlMsg::SetDiagnosticsVisualMarker { .. }
+    )
+}
+
+pub(crate) fn dashboard_control_msg_action(ctrl: &ControlMsg) -> &'static str {
+    match ctrl {
+        ControlMsg::Status { .. } => "status",
+        ControlMsg::Usage => "usage",
+        ControlMsg::Approve { .. } => "approve",
+        ControlMsg::Deny { .. } => "deny",
+        ControlMsg::Skip { .. } => "skip",
+        ControlMsg::ApproveAll { .. } => "approve_all",
+        ControlMsg::AnswerQuestion { .. } => "answer_question",
+        ControlMsg::Input { .. } => "input",
+        ControlMsg::SetAutonomy { .. } => "set_autonomy",
+        ControlMsg::SetApprovalRule { .. } => "set_approval_rule",
+        ControlMsg::SetExternalAgent { .. } => "set_external_agent",
+        ControlMsg::SetCodexCommand { .. } => "set_codex_command",
+        ControlMsg::SetCodexManagedCommand { .. } => "set_codex_managed_command",
+        ControlMsg::SetCodexSandbox { .. } => "set_codex_sandbox",
+        ControlMsg::SetCodexApprovalPolicy { .. } => "set_codex_approval_policy",
+        ControlMsg::SetCodexModel { .. } => "set_codex_model",
+        ControlMsg::SetCodexReasoningEffort { .. } => "set_codex_reasoning_effort",
+        ControlMsg::SetCodexServiceTier { .. } => "set_codex_service_tier",
+        ControlMsg::SetCodexWebSearch { .. } => "set_codex_web_search",
+        ControlMsg::SetCodexNetworkAccess { .. } => "set_codex_network_access",
+        ControlMsg::SetCodexWritableRoots { .. } => "set_codex_writable_roots",
+        ControlMsg::SetCodexManagedContext { .. } => "set_codex_managed_context",
+        ControlMsg::SetCodexContextArchive { .. } => "set_codex_context_archive",
+        ControlMsg::CodexThreadAction { .. } => "codex_thread_action",
+        ControlMsg::RenameSession { .. } => "rename_session",
+        ControlMsg::ConfigureSessionAgent { .. } => "configure_session_agent",
+        ControlMsg::StopSession { .. } => "stop_session",
+        ControlMsg::RestartSession { .. } => "restart_session",
+        ControlMsg::ResumeSession { .. } => "resume_session",
+        ControlMsg::SetClaudeModel { .. } => "set_claude_model",
+        ControlMsg::SetClaudePermissionMode { .. } => "set_claude_permission_mode",
+        ControlMsg::SetClaudeAllowedTools { .. } => "set_claude_allowed_tools",
+        ControlMsg::SetVerbosity { .. } => "set_verbosity",
+        ControlMsg::ScheduleControllerRestart { .. } => "schedule_controller_restart",
+        ControlMsg::ControllerTurnComplete { .. } => "controller_turn_complete",
+        ControlMsg::GetRestartStatus => "get_restart_status",
+        ControlMsg::CancelControllerRestart { .. } => "cancel_controller_restart",
+        ControlMsg::RequestControllerLoopHalt { .. } => "request_controller_loop_halt",
+        ControlMsg::ClearControllerLoopHalt => "clear_controller_loop_halt",
+        ControlMsg::InterveneControllerLoop { .. } => "intervene_controller_loop",
+        ControlMsg::GetControllerLoopStatus => "get_controller_loop_status",
+        ControlMsg::CreateSession { .. } => "create_session",
+        ControlMsg::SpawnSubAgent { .. } => "spawn_sub_agent",
+        ControlMsg::StartTask { .. } => "start_task",
+        ControlMsg::FollowUp { .. } => "follow_up",
+        ControlMsg::CancelFollowUp { .. } => "cancel_follow_up",
+        ControlMsg::EditUserMessage { .. } => "edit_user_message",
+        ControlMsg::QueryDetail { .. } => "query_detail",
+        ControlMsg::RecallMemory { .. } => "recall_memory",
+        ControlMsg::TakeDisplay { .. } => "take_display",
+        ControlMsg::ReleaseDisplay { .. } => "release_display",
+        ControlMsg::GrantUserDisplay { .. } => "grant_user_display",
+        ControlMsg::RevokeUserDisplay { .. } => "revoke_user_display",
+        ControlMsg::CreateBrowserWorkspace { .. } => "create_browser_workspace",
+        ControlMsg::CloseBrowserWorkspace { .. } => "close_browser_workspace",
+        ControlMsg::AcquireBrowserWorkspace { .. } => "acquire_browser_workspace",
+        ControlMsg::ReleaseBrowserWorkspace { .. } => "release_browser_workspace",
+        ControlMsg::ListDisplays => "list_displays",
+        ControlMsg::InvokeSkill { .. } => "invoke_skill",
+        ControlMsg::Quit => "quit",
+        ControlMsg::SetupDebugScreen => "setup_debug_screen",
+        ControlMsg::TeardownDebugScreen => "teardown_debug_screen",
+        ControlMsg::StartDebugRecording => "start_debug_recording",
+        ControlMsg::StopDebugRecording => "stop_debug_recording",
+        ControlMsg::StartRecording { .. } => "start_recording",
+        ControlMsg::StopRecording { .. } => "stop_recording",
+        ControlMsg::DeleteRecording { .. } => "delete_recording",
+        ControlMsg::Interrupt { .. } => "interrupt",
+        ControlMsg::Steer { .. } => "steer",
+        ControlMsg::CancelSteer { .. } => "cancel_steer",
+        ControlMsg::WebRtcSignal { .. } => "webrtc_signal",
+        ControlMsg::PeerFileTransferSignal { .. } => "peer_file_transfer_signal",
+        ControlMsg::PeerDashboardControlSignal { .. } => "peer_dashboard_control_signal",
+        ControlMsg::RequestDisplayInputAuthority { .. } => "request_display_input_authority",
+        ControlMsg::ReleaseDisplayInputAuthority { .. } => "release_display_input_authority",
+        ControlMsg::SetDiagnosticsVisualMarker { .. } => "set_diagnostics_visual_marker",
+    }
+}
+
+pub(crate) async fn api_api_keys_save_response(
+    id: String,
+    params: Option<&serde_json::Value>,
+) -> serde_json::Value {
+    let body_text = params_body_text(params);
+    http_body_response(
+        id,
+        200,
+        crate::web_gateway::handle_set_api_keys(&body_text),
+        "api keys save",
+    )
+}
+
+pub(crate) async fn api_peer_add_response(
+    id: String,
+    params: Option<&serde_json::Value>,
+    runtime: &ControlRuntime,
+) -> serde_json::Value {
+    let Some(registry) = runtime.peer_registry.as_ref() else {
+        return peer_registry_unavailable_response(id);
+    };
+    let body_text = params_body_text(params);
+    let (status, body) =
+        crate::web_gateway::peers_add(registry, runtime.project_root.as_deref(), &body_text).await;
+    http_body_response(id, status, body, "peer add")
+}
+
+pub(crate) async fn api_peer_remove_response(
+    id: String,
+    params: Option<&serde_json::Value>,
+    runtime: &ControlRuntime,
+) -> serde_json::Value {
+    let Some(registry) = runtime.peer_registry.as_ref() else {
+        return peer_registry_unavailable_response(id);
+    };
+    let body_text = params_body_text(params);
+    let (status, body) = crate::web_gateway::peers_remove(registry, &body_text).await;
+    http_body_response(id, status, body, "peer remove")
+}
+
+pub(crate) async fn api_peer_eligible_response(
+    id: String,
+    params: Option<&serde_json::Value>,
+    runtime: &ControlRuntime,
+) -> serde_json::Value {
+    let Some(registry) = runtime.peer_registry.as_ref() else {
+        return peer_registry_unavailable_response(id);
+    };
+    let params = params.cloned().unwrap_or_else(|| serde_json::json!({}));
+    let query = control_capability_query(&params);
+    let (status, body) = crate::web_gateway::peers_eligible(registry, &query);
+    http_body_response(id, status, body, "eligible peers")
+}
+
+pub(crate) async fn api_peer_message_response(
+    id: String,
+    params: Option<&serde_json::Value>,
+    runtime: &ControlRuntime,
+) -> serde_json::Value {
+    let Some(registry) = runtime.peer_registry.as_ref() else {
+        return peer_registry_unavailable_response(id);
+    };
+    let params = params.cloned().unwrap_or_else(|| serde_json::json!({}));
+    let peer_id = string_param(&params, &["peer_id", "peerId", "host_id", "hostId", "id"]);
+    if peer_id.is_empty() {
+        return missing_param_response(id, "peer_id");
+    }
+    let body_text = serde_json::to_string(&params).unwrap_or_else(|_| "{}".to_string());
+    let (status, body) =
+        crate::web_gateway::peers_send_message(registry, &peer_id, &body_text).await;
+    http_body_response(id, status, body, "peer message")
+}
+
+pub(crate) async fn api_peer_task_response(
+    id: String,
+    params: Option<&serde_json::Value>,
+    runtime: &ControlRuntime,
+) -> serde_json::Value {
+    let Some(registry) = runtime.peer_registry.as_ref() else {
+        return peer_registry_unavailable_response(id);
+    };
+    let params = params.cloned().unwrap_or_else(|| serde_json::json!({}));
+    let peer_id = string_param(&params, &["peer_id", "peerId", "host_id", "hostId", "id"]);
+    if peer_id.is_empty() {
+        return missing_param_response(id, "peer_id");
+    }
+    let body_text = serde_json::to_string(&params).unwrap_or_else(|_| "{}".to_string());
+    let (status, body) =
+        crate::web_gateway::peers_delegate_task(registry, &peer_id, &body_text).await;
+    http_body_response(id, status, body, "peer task")
+}
+
+pub(crate) async fn api_peer_approval_response(
+    id: String,
+    params: Option<&serde_json::Value>,
+    runtime: &ControlRuntime,
+) -> serde_json::Value {
+    let Some(registry) = runtime.peer_registry.as_ref() else {
+        return peer_registry_unavailable_response(id);
+    };
+    let params = params.cloned().unwrap_or_else(|| serde_json::json!({}));
+    let peer_id = string_param(&params, &["peer_id", "peerId", "host_id", "hostId", "id"]);
+    if peer_id.is_empty() {
+        return missing_param_response(id, "peer_id");
+    }
+    let body_text = serde_json::to_string(&params).unwrap_or_else(|_| "{}".to_string());
+    let (status, body) =
+        crate::web_gateway::peers_resolve_approval(registry, &peer_id, &body_text).await;
+    http_body_response(id, status, body, "peer approval")
+}
+
+pub(crate) async fn api_peer_webrtc_signal_response(
+    id: String,
+    params: Option<&serde_json::Value>,
+    runtime: &ControlRuntime,
+) -> serde_json::Value {
+    let Some(registry) = runtime.peer_registry.as_ref() else {
+        return peer_registry_unavailable_response(id);
+    };
+    let params = params.cloned().unwrap_or_else(|| serde_json::json!({}));
+    let peer_id = string_param(&params, &["peer_id", "peerId", "host_id", "hostId", "id"]);
+    if peer_id.is_empty() {
+        return missing_param_response(id, "peer_id");
+    }
+    let body_text = serde_json::to_string(&params).unwrap_or_else(|_| "{}".to_string());
+    let (status, body) =
+        crate::web_gateway::peers_webrtc_signal(registry, &peer_id, &body_text, &runtime.bus).await;
+    http_body_response(id, status, body, "peer webrtc signal")
+}
+
+pub(crate) async fn api_peer_file_transfer_signal_response(
+    id: String,
+    params: Option<&serde_json::Value>,
+    runtime: &ControlRuntime,
+) -> serde_json::Value {
+    let Some(registry) = runtime.peer_registry.as_ref() else {
+        return peer_registry_unavailable_response(id);
+    };
+    let params = params.cloned().unwrap_or_else(|| serde_json::json!({}));
+    let peer_id = string_param(&params, &["peer_id", "peerId", "host_id", "hostId", "id"]);
+    if peer_id.is_empty() {
+        return missing_param_response(id, "peer_id");
+    }
+    let body_text = serde_json::to_string(&params).unwrap_or_else(|_| "{}".to_string());
+    let (status, body) = crate::web_gateway::peers_file_transfer_signal(
+        registry,
+        &peer_id,
+        &body_text,
+        &runtime.bus,
+    )
+    .await;
+    http_body_response(id, status, body, "peer file-transfer signal")
+}
+
+pub(crate) async fn api_peer_dashboard_control_signal_response(
+    id: String,
+    params: Option<&serde_json::Value>,
+    runtime: &ControlRuntime,
+) -> serde_json::Value {
+    let Some(registry) = runtime.peer_registry.as_ref() else {
+        return peer_registry_unavailable_response(id);
+    };
+    let params = params.cloned().unwrap_or_else(|| serde_json::json!({}));
+    let peer_id = string_param(&params, &["peer_id", "peerId", "host_id", "hostId", "id"]);
+    if peer_id.is_empty() {
+        return missing_param_response(id, "peer_id");
+    }
+    let body_text = serde_json::to_string(&params).unwrap_or_else(|_| "{}".to_string());
+    let (status, body) = crate::web_gateway::peers_dashboard_control_signal(
+        registry,
+        &peer_id,
+        &body_text,
+        &runtime.bus,
+    )
+    .await;
+    http_body_response(id, status, body, "peer dashboard-control signal")
+}
+
+pub(crate) async fn api_peer_pairing_invite_response(
+    id: String,
+    params: Option<&serde_json::Value>,
+) -> serde_json::Value {
+    let body_text = params_body_text(params);
+    let (status, body) = crate::web_gateway::peers_pairing_invite(&body_text);
+    http_body_response(id, status, body, "peer pairing invite")
+}
+
+pub(crate) async fn api_peer_pairing_join_response(
+    id: String,
+    params: Option<&serde_json::Value>,
+    runtime: &ControlRuntime,
+) -> serde_json::Value {
+    let Some(registry) = runtime.peer_registry.as_ref() else {
+        return peer_registry_unavailable_response(id);
+    };
+    let body_text = params_body_text(params);
+    let (status, body) = crate::web_gateway::peers_pairing_join(
+        registry,
+        runtime.project_root.as_deref(),
+        &body_text,
+    )
+    .await;
+    http_body_response(id, status, body, "peer pairing join")
+}
+
+pub(crate) async fn api_peer_pairing_request_access_response(
+    id: String,
+    params: Option<&serde_json::Value>,
+) -> serde_json::Value {
+    let body_text = params_body_text(params);
+    let (status, body) = crate::web_gateway::peers_pairing_request_access(&body_text).await;
+    http_body_response(id, status, body, "peer access request")
+}
+
+pub(crate) async fn api_peer_pairing_request_access_poll_response(
+    id: String,
+    params: Option<&serde_json::Value>,
+    runtime: &ControlRuntime,
+) -> serde_json::Value {
+    let body_text = params_body_text(params);
+    let (status, body) = crate::web_gateway::peers_pairing_request_access_poll(
+        runtime.peer_registry.as_ref(),
+        runtime.project_root.as_deref(),
+        &body_text,
+    )
+    .await;
+    http_body_response(id, status, body, "peer access request poll")
+}
+
+pub(crate) async fn api_peer_pairing_requests_response(id: String) -> serde_json::Value {
+    let (status, body) = crate::web_gateway::peers_pairing_requests_list();
+    http_body_response(id, status, body, "peer access requests")
+}
+
+pub(crate) async fn api_peer_pairing_request_decision_response(
+    id: String,
+    params: Option<&serde_json::Value>,
+) -> serde_json::Value {
+    let params = params.cloned().unwrap_or_else(|| serde_json::json!({}));
+    let request_id = string_param(&params, &["request_id", "requestId", "code", "id"]);
+    if request_id.is_empty() {
+        return missing_param_response(id, "request_id");
+    }
+    let op = string_param(&params, &["op", "decision", "action"]);
+    let op = if op.is_empty() {
+        "approve".to_string()
+    } else {
+        op
+    };
+    let body_text = serde_json::to_string(&params).unwrap_or_else(|_| "{}".to_string());
+    let (status, body) =
+        crate::web_gateway::peers_pairing_request_decision(&request_id, &op, &body_text);
+    http_body_response(id, status, body, "peer access request decision")
+}
+
+pub(crate) async fn api_peer_pairing_identities_response(id: String) -> serde_json::Value {
+    let (status, body) = crate::web_gateway::peers_pairing_identities_list();
+    http_body_response(id, status, body, "peer identities")
+}
+
+pub(crate) async fn api_peer_pairing_identity_revoke_response(
+    id: String,
+    params: Option<&serde_json::Value>,
+) -> serde_json::Value {
+    let body_text = params_body_text(params);
+    let (status, body) = crate::web_gateway::peers_pairing_identity_revoke(&body_text);
+    http_body_response(id, status, body, "peer identity revoke")
+}
+
+pub(crate) async fn api_coordinator_route_response(
+    id: String,
+    params: Option<&serde_json::Value>,
+    runtime: &ControlRuntime,
+) -> serde_json::Value {
+    let Some(registry) = runtime.peer_registry.as_ref() else {
+        return peer_registry_unavailable_response(id);
+    };
+    let body_text = params_body_text(params);
+    let (status, body) = crate::web_gateway::coordinator_route(registry, &body_text).await;
+    http_body_response(id, status, body, "coordinator route")
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::*;
+    use crate::dashboard_control::tests::{runtime};
+
+    struct DashboardControlStubDisplayBackend;
+
+    #[async_trait::async_trait]
+    impl crate::display::DisplayBackend for DashboardControlStubDisplayBackend {
+        async fn start_capture(
+            &self,
+            _fps: u32,
+        ) -> Result<mpsc::Receiver<crate::display::Frame>, CallerError> {
+            let (_tx, rx) = mpsc::channel(1);
+            Ok(rx)
+        }
+
+        async fn stop_capture(&self) {}
+
+        async fn inject_input(
+            &self,
+            _event: crate::display::InputEvent,
+        ) -> Result<(), CallerError> {
+            Ok(())
+        }
+
+        fn resolution(&self) -> (u32, u32) {
+            (64, 64)
+        }
+
+        fn kind(&self) -> &'static str {
+            "dashboard-control-stub"
+        }
+    }
+
+    #[tokio::test]
+    async fn api_voice_session_preserves_endpoint_error_metadata() {
+        let mut rt = runtime();
+        rt.config = serde_json::json!({
+            "provider": "unsupported-voice-provider",
+            "model": "unused",
+        });
+        let response = api_voice_session_response("voice1".to_string(), &rt).await;
+        assert_eq!(response["t"], "response");
+        assert_eq!(response["id"], "voice1");
+        assert_eq!(response["ok"], true);
+        assert_eq!(response["result"]["_httpStatus"], 502);
+        assert_eq!(response["result"]["_httpOk"], false);
+        assert_eq!(
+            response["result"]["error"],
+            "Unknown provider: unsupported-voice-provider"
+        );
+    }
+
+    #[tokio::test]
+    async fn api_mcp_tool_call_reports_unavailable_server_as_http_error() {
+        let rt = runtime();
+        let response = api_mcp_tool_call_response(
+            "mcp1".to_string(),
+            Some(&serde_json::json!({
+                "mcp_id": 7,
+                "session_id": "session-1",
+                "name": "get_status",
+            })),
+            &rt,
+        )
+        .await;
+        assert_eq!(response["t"], "response");
+        assert_eq!(response["id"], "mcp1");
+        assert_eq!(response["ok"], true);
+        assert_eq!(response["result"]["_httpStatus"], 503);
+        assert_eq!(response["result"]["_httpOk"], false);
+        assert_eq!(response["result"]["id"], 7);
+        assert_eq!(response["result"]["error"]["code"], -32603);
+        assert_eq!(
+            response["result"]["error"]["message"],
+            "MCP server not available"
+        );
+    }
+
+    #[tokio::test]
+    async fn api_control_msg_dispatches_allowlisted_settings_actions_only() {
+        let rt = runtime();
+        let mut events = rt.bus.subscribe();
+        let response = api_control_msg_response(
+            "ctrl1".to_string(),
+            Some(&serde_json::json!({
+                "message": {
+                    "action": "set_codex_sandbox",
+                    "mode": "workspace-write",
+                }
+            })),
+            &rt,
+        )
+        .await;
+        assert_eq!(response["t"], "response");
+        assert_eq!(response["ok"], true);
+        assert_eq!(response["result"]["ok"], true);
+        assert_eq!(response["result"]["action"], "set_codex_sandbox");
+
+        let mut saw_control = false;
+        for _ in 0..4 {
+            let event = tokio::time::timeout(Duration::from_secs(1), events.recv())
+                .await
+                .unwrap()
+                .unwrap();
+            if let AppEvent::ControlCommand(ControlMsg::SetCodexSandbox { mode }) = event {
+                assert_eq!(mode, "workspace-write");
+                saw_control = true;
+                break;
+            }
+        }
+        assert!(saw_control, "allowed control message did not reach the bus");
+
+        let rejected = api_control_msg_response(
+            "ctrl2".to_string(),
+            Some(&serde_json::json!({
+                "message": {
+                    "action": "create_session",
+                    "task": "do something",
+                }
+            })),
+            &rt,
+        )
+        .await;
+        assert_eq!(rejected["t"], "response");
+        assert_eq!(rejected["ok"], true);
+        assert_eq!(rejected["result"]["ok"], false);
+        assert_eq!(rejected["result"]["_httpStatus"], 400);
+        assert!(rejected["result"]["error"]
+            .as_str()
+            .unwrap_or("")
+            .contains("not available over dashboard WebRTC"));
+    }
+
+    #[tokio::test]
+    async fn api_session_control_msg_dispatches_lifecycle_actions_only() {
+        let rt = runtime();
+        let mut events = rt.bus.subscribe();
+        let response = api_session_control_msg_response(
+            "session-ctrl1".to_string(),
+            Some(&serde_json::json!({
+                "message": {
+                    "action": "interrupt",
+                    "session_id": "session-a",
+                }
+            })),
+            &rt,
+        )
+        .await;
+        assert_eq!(response["t"], "response");
+        assert_eq!(response["ok"], true);
+        assert_eq!(response["result"]["ok"], true);
+        assert_eq!(response["result"]["action"], "interrupt");
+
+        let mut saw_control = false;
+        for _ in 0..4 {
+            let event = tokio::time::timeout(Duration::from_secs(1), events.recv())
+                .await
+                .unwrap()
+                .unwrap();
+            if let AppEvent::ControlCommand(ControlMsg::Interrupt { session_id, .. }) = event {
+                assert_eq!(session_id.as_deref(), Some("session-a"));
+                saw_control = true;
+                break;
+            }
+        }
+        assert!(saw_control, "session control message did not reach the bus");
+
+        let accepted_create = api_session_control_msg_response(
+            "session-ctrl2".to_string(),
+            Some(&serde_json::json!({
+                "message": {
+                    "action": "create_session",
+                    "task": "noop",
+                }
+            })),
+            &rt,
+        )
+        .await;
+        assert_eq!(accepted_create["t"], "response");
+        assert_eq!(accepted_create["ok"], true);
+        assert_eq!(accepted_create["result"]["ok"], true);
+        assert_eq!(accepted_create["result"]["action"], "create_session");
+
+        let rejected_settings = api_session_control_msg_response(
+            "session-ctrl3".to_string(),
+            Some(&serde_json::json!({
+                "message": {
+                    "action": "set_codex_sandbox",
+                    "mode": "workspace-write",
+                }
+            })),
+            &rt,
+        )
+        .await;
+        assert_eq!(rejected_settings["t"], "response");
+        assert_eq!(rejected_settings["ok"], true);
+        assert_eq!(rejected_settings["result"]["ok"], false);
+        assert_eq!(rejected_settings["result"]["_httpStatus"], 400);
+        assert!(rejected_settings["result"]["error"]
+            .as_str()
+            .unwrap_or("")
+            .contains("not available over dashboard session WebRTC"));
+    }
+
+    #[tokio::test]
+    async fn api_dashboard_action_msg_dispatches_small_dashboard_actions_only() {
+        let rt = runtime();
+        let mut events = rt.bus.subscribe();
+        let response = api_dashboard_action_msg_response(
+            "dash-action1".to_string(),
+            Some(&serde_json::json!({
+                "message": {
+                    "action": "close_browser_workspace",
+                    "workspace_id": "workspace-a",
+                    "reason": "test",
+                }
+            })),
+            &rt,
+        )
+        .await;
+        assert_eq!(response["t"], "response");
+        assert_eq!(response["ok"], true);
+        assert_eq!(response["result"]["ok"], true);
+        assert_eq!(response["result"]["action"], "close_browser_workspace");
+
+        let mut saw_control = false;
+        for _ in 0..4 {
+            let event = tokio::time::timeout(Duration::from_secs(1), events.recv())
+                .await
+                .unwrap()
+                .unwrap();
+            if let AppEvent::ControlCommand(ControlMsg::CloseBrowserWorkspace {
+                workspace_id,
+                ..
+            }) = event
+            {
+                assert_eq!(workspace_id, "workspace-a");
+                saw_control = true;
+                break;
+            }
+        }
+        assert!(
+            saw_control,
+            "dashboard action message did not reach the bus"
+        );
+
+        let accepted_thread = api_dashboard_action_msg_response(
+            "dash-action2".to_string(),
+            Some(&serde_json::json!({
+                "message": {
+                    "action": "codex_thread_action",
+                    "session_id": "session-a",
+                    "op": "new",
+                    "params": {},
+                }
+            })),
+            &rt,
+        )
+        .await;
+        assert_eq!(accepted_thread["t"], "response");
+        assert_eq!(accepted_thread["ok"], true);
+        assert_eq!(accepted_thread["result"]["action"], "codex_thread_action");
+
+        let rejected_settings = api_dashboard_action_msg_response(
+            "dash-action3".to_string(),
+            Some(&serde_json::json!({
+                "message": {
+                    "action": "set_codex_sandbox",
+                    "mode": "workspace-write",
+                }
+            })),
+            &rt,
+        )
+        .await;
+        assert_eq!(rejected_settings["t"], "response");
+        assert_eq!(rejected_settings["ok"], true);
+        assert_eq!(rejected_settings["result"]["ok"], false);
+        assert_eq!(rejected_settings["result"]["_httpStatus"], 400);
+        assert!(rejected_settings["result"]["error"]
+            .as_str()
+            .unwrap_or("")
+            .contains("not available over dashboard action WebRTC"));
+    }
+
+    #[tokio::test]
+    async fn api_diagnostics_visual_freshness_appends_ndjson_batch() {
+        let session_id = format!("dashboard-control-test-vf-{}", std::process::id());
+        if let Some(path) = crate::diagnostics::visual_freshness_path(&session_id) {
+            let _ = std::fs::remove_file(&path);
+        }
+        let ndjson = "{\"t\":\"session_start\"}\n{\"t\":\"summary\"}\n";
+        let response = api_diagnostics_visual_freshness_response(
+            "diag-vf".to_string(),
+            Some(&serde_json::json!({
+                "session_id": session_id.clone(),
+                "body": ndjson,
+            })),
+        )
+        .await;
+        assert_eq!(response["t"], "response");
+        assert_eq!(response["ok"], true);
+        assert_eq!(response["result"]["ok"], true);
+        assert_eq!(response["result"]["_httpStatus"], 200);
+        assert_eq!(response["result"]["written"], ndjson.len());
+
+        let path =
+            crate::diagnostics::visual_freshness_path(&session_id).expect("diagnostics path");
+        let written = std::fs::read_to_string(&path).expect("diagnostics transcript");
+        assert_eq!(written, ndjson);
+        let _ = std::fs::remove_file(path);
+    }
+
+    #[tokio::test]
+    async fn api_dashboard_action_msg_applies_diagnostics_visual_marker_to_display_registry() {
+        let rt = runtime();
+        let registry = Arc::new(tokio::sync::RwLock::new(
+            crate::display::SessionRegistry::new(),
+        ));
+        let display_session = Arc::new(crate::display::DisplaySession::new(
+            2,
+            Arc::new(DashboardControlStubDisplayBackend),
+        ));
+        registry
+            .write()
+            .await
+            .insert(2, Arc::clone(&display_session));
+        {
+            let mut session = rt.shared_session.write().await;
+            session.session_registry = Some(Arc::clone(&registry));
+        }
+
+        let response = api_dashboard_action_msg_response(
+            "dash-action-marker".to_string(),
+            Some(&serde_json::json!({
+                "message": {
+                    "action": "set_diagnostics_visual_marker",
+                    "display_id": 2,
+                    "enabled": true,
+                }
+            })),
+            &rt,
+        )
+        .await;
+        assert_eq!(response["t"], "response");
+        assert_eq!(response["ok"], true);
+        assert_eq!(response["result"]["ok"], true);
+        assert_eq!(
+            response["result"]["action"],
+            "set_diagnostics_visual_marker"
+        );
+        assert_eq!(response["result"]["display_id"], 2);
+        assert_eq!(response["result"]["registry_available"], true);
+        assert_eq!(response["result"]["active_display_updated"], true);
+        assert!(
+            display_session.diagnostics_visual_marker_enabled(),
+            "dashboard-control RPC did not toggle the live display session"
+        );
+    }
+
+    #[tokio::test]
+    async fn peer_webrtc_signal_returns_http_error_metadata() {
+        let (log_tx, _log_rx) =
+            tokio::sync::mpsc::channel::<crate::peer::event::TaggedPeerEvent>(8);
+        let mut rt = runtime();
+        rt.peer_registry = Some(crate::peer::PeerRegistry::new(log_tx));
+
+        let params = serde_json::json!({
+            "peer_id": "missing-peer",
+            "display_id": 0,
+            "session_id": "dashboard-test-session",
+            "signal": { "kind": "close" },
+        });
+        let response =
+            api_peer_webrtc_signal_response("webrtc1".to_string(), Some(&params), &rt).await;
+
+        assert_eq!(response["t"], "response");
+        assert_eq!(response["ok"], true);
+        assert_eq!(response["result"]["_httpOk"], false);
+        assert_eq!(response["result"]["_httpStatus"], 404);
+        assert_eq!(response["result"]["error"], "peer not found");
+    }
+
+    #[tokio::test]
+    async fn state_snapshot_rpc_returns_bootstrap_message_shape() {
+        let rt = runtime();
+        let snapshot = api_state_snapshot_response("snap1".to_string(), &rt).await;
+        assert_eq!(snapshot["t"], "response");
+        assert_eq!(snapshot["id"], "snap1");
+        assert_eq!(snapshot["ok"], true);
+        assert_eq!(snapshot["result"]["t"], "state_snapshot");
+        assert_eq!(snapshot["result"]["connection_id"], "session-1");
+        assert_eq!(snapshot["result"]["config"]["provider"], "openai");
+        assert_eq!(snapshot["result"]["session_id"], "");
+        assert!(snapshot["result"]["state"].is_object());
+    }
+
+    #[tokio::test]
+    async fn session_log_replay_rpc_returns_empty_replay_without_active_log() {
+        let rt = runtime();
+        let replay = api_session_log_replay_response("replay1".to_string(), &rt).await;
+        assert_eq!(replay["t"], "response");
+        assert_eq!(replay["id"], "replay1");
+        assert_eq!(replay["ok"], true);
+        assert_eq!(replay["result"]["t"], "log_replay");
+        assert_eq!(replay["result"]["available"], false);
+        assert_eq!(replay["result"]["entries"].as_array().unwrap().len(), 0);
+    }
+
+    #[tokio::test]
+    async fn display_bootstrap_rpc_returns_empty_frames_without_active_displays() {
+        let rt = runtime();
+        let bootstrap = api_display_bootstrap_response("disp1".to_string(), &rt).await;
+        assert_eq!(bootstrap["t"], "response");
+        assert_eq!(bootstrap["id"], "disp1");
+        assert_eq!(bootstrap["ok"], true);
+        assert_eq!(bootstrap["result"]["frame_count"], 0);
+        assert_eq!(bootstrap["result"]["frames"].as_array().unwrap().len(), 0);
+        assert!(bootstrap["result"]["omitted"]
+            .as_array()
+            .unwrap()
+            .iter()
+            .any(|value| value.as_str() == Some("display_input_authority_state")));
+    }
+
+    #[tokio::test]
+    async fn display_webrtc_signal_rpc_reports_missing_display() {
+        let rt = runtime();
+        let params = serde_json::json!({
+            "signal": "offer",
+            "display_id": 99,
+            "sdp": "synthetic-offer",
+        });
+        let response =
+            api_display_webrtc_signal_response("sig1".to_string(), Some(&params), &rt).await;
+        assert_eq!(response["t"], "response");
+        assert_eq!(response["id"], "sig1");
+        assert_eq!(response["ok"], false);
+        assert_eq!(response["status"], 404);
+        assert_eq!(response["display_id"], 99);
+        assert_eq!(response["error"], "display session not found");
+    }
+
+    #[tokio::test]
+    async fn external_session_activity_replay_rpc_returns_empty_frames_without_attached_sessions() {
+        let rt = runtime();
+        let replay = api_external_session_activity_replay_response("ext1".to_string(), &rt).await;
+        assert_eq!(replay["t"], "response");
+        assert_eq!(replay["id"], "ext1");
+        assert_eq!(replay["ok"], true);
+        assert_eq!(replay["result"]["frame_count"], 0);
+        assert_eq!(replay["result"]["frames"].as_array().unwrap().len(), 0);
+    }
+
+    #[tokio::test]
+    async fn dashboard_bootstrap_rpc_returns_ordered_bootstrap_frames() {
+        let rt = runtime();
+        let bootstrap = api_dashboard_bootstrap_response("boot1".to_string(), &rt).await;
+        assert_eq!(bootstrap["t"], "response");
+        assert_eq!(bootstrap["id"], "boot1");
+        assert_eq!(bootstrap["ok"], true);
+        let frames = bootstrap["result"]["frames"].as_array().unwrap();
+        assert_eq!(bootstrap["result"]["frame_count"], frames.len());
+        assert_eq!(frames[0]["t"], "state_snapshot");
+        assert_eq!(frames[1]["t"], "browser_workspace_snapshot");
+        assert_eq!(frames[2]["t"], "log_replay");
+        assert!(!bootstrap["result"]["omitted"]
+            .as_array()
+            .unwrap()
+            .iter()
+            .any(|value| value.as_str() == Some("display_ready")));
+        assert!(bootstrap["result"]["omitted"]
+            .as_array()
+            .unwrap()
+            .iter()
+            .any(|value| value.as_str() == Some("display_input_authority_state")));
+        assert!(!bootstrap["result"]["omitted"]
+            .as_array()
+            .unwrap()
+            .iter()
+            .any(|value| value.as_str() == Some("external_session_activity_replay")));
+    }
+
+    #[tokio::test]
+    async fn worktree_rpcs_preserve_cache_and_error_status() {
+        let rt = runtime();
+        {
+            let mut cache = rt.worktree_inventory_cache.lock().unwrap();
+            *cache = Some(
+                serde_json::json!({
+                    "worktrees": [{ "path": "/tmp/wt", "branch": "feature" }],
+                    "summary": { "worktrees": 1 },
+                })
+                .to_string(),
+            );
+        }
+
+        let cached = api_worktrees_response("wt1".to_string(), &rt).await;
+        assert_eq!(cached["t"], "response");
+        assert_eq!(cached["ok"], true);
+        assert_eq!(cached["result"]["summary"]["worktrees"], 1);
+
+        let invalid_remove =
+            api_worktrees_remove_response("wt2".to_string(), Some(&serde_json::json!({})), &rt)
+                .await;
+        assert_eq!(invalid_remove["t"], "response");
+        assert_eq!(invalid_remove["ok"], true);
+        assert_eq!(invalid_remove["result"]["ok"], false);
+        assert_eq!(invalid_remove["result"]["_httpStatus"], 400);
+        assert_eq!(invalid_remove["result"]["_httpOk"], false);
+    }
+}
